Silence compiler warnings

+++ src/genwiki.c
@@ -37,7 +37,7 @@ char * sanitize_link(const char * doi) {
   int i, j, MAXLEN=4096;
   char buf[MAXLEN], *r;
 
-  for(i=0, j=0; i<strlen(doi); i++) {
+  for(i=0, j=0; i<(int) strlen(doi); i++) {
     switch(doi[i]) {
     case('!'):
       buf[j++]='%';
@@ -217,7 +217,7 @@ int main(void) {
 	  /* Get functional keyword */
 	  fname = xc_functional_get_name(func_id);
 	  /* Convert functional keyword to upper case */
-	  for(i=0; i<strlen(fname); i++)
+	  for(i=0; i<(int) strlen(fname); i++)
 	    fname[i]=toupper(fname[i]);
 
 	  /* Print out info */
